refactor: move emissions public headers out of versioned include path
Aligns with the library convention where all public headers live directly
under include/mechanism_configuration/ without a version subdirectory.
The v1 version is still expressed in the namespace and source layout.
